First commit of AngularJS Project: OLL + PLL Cases
Hopefully will progress to both OLL and PLL cases, but for now, I'm probably just happy to have at least the OLL cases. Plus I already have all of the framework basically set in place such that I could probably just use the same code for a different page, but with a different set of cases.
If possible, the next steps will include cleannig up my AngularJS code so that I can prepare it for unit testing. I think that the unit testing will dictate some of the structure of the code, so I'm a bit apprehensive about doing this all at once; however, the code is frankly way too nasty for me to even really think about looking at it and updating it regularly.
After it is testable, then I might add the PLL cases, altohugh I could just also record all of the data say, tonight, and then continue working on unit testing tomorrow.
I've decided to commit today because if I'm going to be messing around with the code, I need to know that I have a 100% working version at least. Without something like Git, this would be 100% impossible. Most of my time on my laptop, including 17 hours and 41 minutes on Eclipse and about 5 hours on Safari (I already accounted for the time that I spent watching Criminal Minds on Safari), for the past three days, has been on this application, and for all 17 of those hours to go to waste would be devastating.
Git also gives me a nice continuity feeling, that I have all these different versions of something, like a baby photo album, with all of the baby's firsts....except for an application.
So the next step will be cleaning up the code, tomorrow, Too late to start such an endeavor now.