Created the cookieSavedData module
This will hopefully replace the very bulky code that I currently have in the ContentController controller module. cookieSavedData is a factory service that is able to get, set, and remove data from the factory, which holds all of the data for flash cards, and I really should rename this to the flashCardData factory module.
Other methods included this factory are wrappers for saving all of the flashcard data in cookies.