Integrated the new .js files with the HTML
Everything seems to be working, i.e. the factory's cookie methods save data, I've interwoven the methods of the factory in so that I can hopefully just end up using just the gfactory, instead of having to attach things to the factory, andthat's been good.
There were several bugs with trying to integrate the modules with the HTML because there were several forgetful errors on my part. I'm trying to get this code to comply with normal AngularJS standards so in the midst of trying to make the code less confusing, I was confused myself. I might move on to unit testing now because I actually have some testable modules.