Forgot about one of the other modules - Timer
Had to make a file for the Timer module. Because Timer is one of the smaller modules, I didn't take that long, so I just went ahead and reformatted it such that it complies with the style guide. It's a short enough module, like Scrambler, that it doesn't take too long, thankfully.
Also added another folder for tests, but I think I might get rid of it and have the tests alongside the actual stuff. But I'm still trying to research what a good format for that is.