Everything is broken
I think it may be partially that the cookies were all destroyed but also that I changed the format of things so I'm going to go back to the future / back in time and see if that has any effect, because I made this system so that it would work even when there Aren't Any Cookies because that would just not make sense
I think that I might also try to learn sql and make a database instead because honestly this is a nightmare. If going back in time doesn't help, that's what I'll do.