Fixed everything (?)
I don't know how, but everything suddenly began working again, with respect to the cookies. I think what really happened was that I had a sufficient number of cookies for it to be happy, and it just started working again almost on its own. So I'm going to have to make this more foolproof at some point, preferably soon.
I also implemented a statistics factory, as well as a time conversion factory + filter and a new table specifically for statistics, except it wasn't working as a directive for some reason, so that kind of sucks, but whatever.
Point is, it finally works again, and I just have to be careful about testing.