Added some preventative medicine so to speak for cookies
I've put already-initiated values into the cookieData value, in hopes of remedying cases where the user doesn't already have a cookie for a given object and everything blows up in a giant ball of fire, and and these values are really to make sure that the arrays and objects are at least somewhat defined. I haven't really tested this out, however.