Added factories and filters
Added a statistics factory, a time conversion factory, and I'm currently working on a timer records factory that will organize the times by the date they were recorded on. I tried making a time stats directive, but for some reason, it never worked.
I also made a filter that uses the time conversion factory to make a time that is in the format of milliseconds into a readable format.