Added all of the test declarations for statistics factory, starting on flashcards controller
It took me at least a solid hour to write in all of the tests that I'll be using for the Statistics factory, and I didn't even flesh any of them out, except for the 'is this thing defined' unit tests.
I'm about to start on giving more tests to the flash card controller now. I'll keep writing tests later. I'm just glad that this whole thing works.
In the meantime, I put 'expect(true).toBe(false)' in-between the brackets for all of the incomplete tests, so that I know how many tests actually work.