Finished 75% of the flash-cards factory tests
Nearly 75-80%, or even 85%, done with the flash cards factory tests. I don't even use most of this library, but only because I didn't really feel like trying to use it in my code because I couldn't really ensure that it would work the same as my actual code. If I do this correctly, I might be able to just call saves in the functions that also set them, so that I could have less watches, and therefore free up some computational power.
So anyway, the unit testing is going well...sort of. I can never get coverage to actually do anything in the command line, and it only makes a html output every so often. So, anyway, I'm going to continue testing...