Reorganized my app files, the html editor broke, finished 84% flash cards UTs
Anyway, approximately 84% of my flash card unit tests, or at least I did about an hour ago.
The organization of the app files were confusing vs the way that the files are actually displayed by the coverage plugin.
I want to start working on the statistics unit tests. I think I'm going to be able to simplify things further for myself for this particular set of unit tests...There's wayy too many right now.
Also, I had to switch html editors because for some reason, the one I used to use started throwing a null pointer exception in Java, so that was fun.