Fiddled around with Bamboo, really unsatisfied with code coverage
Not exactly with the code coverage itself, but with the visual
representation of it. I found a really nice example of what the visual
representation of the tests could look like, but it doesn't appear to
work for some reason, unfortunately.
Bamboo also feels...really impenetrable. But anyway, I'm going to try to
push tonight and see what happens.