Fixed bug, redesigned timer controller's records-keeping
The way I used to have the records saved was in just one cookie, which
got way too messy once I started doing...other things...
Basically, the files were too large to be saved as cookies anymore. This
also probably means that I'm going to have to redo a bunch of my
tests...at least I wasn't even really started on the timer controller's
tests...and I made it such that the records get converted in other
places to the original format so that I don't have to cry.

However, I also had to redo the way I create a .csv document, which
meant that I introduced a bug where it wouldn't read the file properly.
What happened was that the string "data:text/csv;charset=utf-8" was not
at the beginning anymore because right after that declaration, I had a
line that reset the variable again, using only a "=" instead of "+=".