Finally made a task to clean out my reports
This task helps me by automatically cleaning out my tests every time I
call 'grunt clean'. I might add this to the grunt task test, so that it
happens every single time I actually test.
This also means that the tests are always going to clean up that folder.

This way, I might be able to get Bamboo to actually find the test
output, because it has been really sketchy so far to actually find it.