Karma acting up, Resolved failing tests 23 -> 2
Now, the only tests that are failing are ones that I never really
resolved in the first place. Hurray!
But in other news, Karma or Grunt or both is acting up. There's so much
being printed out to log, I feel like it must be using logLevel:
"config.LOG_DEBUG", but it doesn't appear to be, and when I try to
change this, it then started saying that it was using the log debug,
even though I specified ".LOG_INFO". This is with "grunt test", not with
"grunt buildtest". Either way, this behavior is extremely weird.