Got the last two tests working
Turns out all I had to do was change the definition of a function. I decided that I should really just offer support for both types of data, just in case. In the meantime, I also really need to update it such that you can have both for all of the functions, which really should just mean copying and pasting or something...