Added a lot of CSS/style changes.
I think that the next thing that I want to do is to change all of my CSS
files into LESS files, so that I can hopefully have LESS confusion while
looking at the CSS. I'll also just be able to add the less-to-css
converter to my grunt setup, and maybe even add the minifier for JS and
CSS, because honestly this is getting to be way too many files.
LESS should make it 10x easier for me to change the look and feel of my
webpages, so very excited on that note.