Added animations + own accordion system, plus more
The only thing missing in the animations for the accordion is getting
the whole thing to actually render the left navigation bar properly.
I've got the logic in there, but it doesn't seem to be working. It's
supposed to bring up the whole left bar as it's animating the slide
down, or bring down the whole left bar as it's animating the slide up,
but while it does this, it doesn't seem to do enough of it. I think part
of the problem is that I have a few nested <dividing-box> tags. Maybe I
should come up with a version where you can have inner <dividing-box>'s
and still be able to animate?

Tooltips (using Bootstrap) were also implemented. I might look more into
Bootstrap, to see if it has any other tools that I could possibly use.

I also am looking into LESS implementation, as well as still
investigating React. Additional support for alternate formats, as well
as different settings, are also things that I want to implement.

I may possibly even try to set up a real server for this, instead of
whatever I have going on right now.

I should also make either an "about" page or a "how to use" page -- or
both.

I also want to begin implementing the PLL side of this project soon.