Tweaks
Got rid of the trailing comma on the times records. They should now be
printed out in the right order, which is good.

I'm planning on branching this out so that I can fix it such that I can
have the PLL cases in there by just passing a variable to a factory.