Changed up pllCases, adedd ContentControl factory, etc.
Added solve1: {} and solve2: {} to the pllCases value by using RegEx's
and atom's beautiful find and replace, which although it isn't quite
like the Eclipse find and replace, holds its own.

I'm also converting the ContentController into a factory, and then
I'll have a PLL controller and a OLL controller that will take care of
the discrepancy between the two, such as making sure it's using the
right part of the the cookieStrings, which now has two sub-objects, pll
and oll, which have strings prefixed OLL and PLL, respectively, of
course.

My next step will be to find the images for the PLL content, and to
implement the CaseManager a bit further. I think ContentController could
be abstracted a little better too.