Moved oll images, added pll images, tweakes pll/ollCases
I finally added the image urls to the pll and oll cases values, using
regexes!!! Very exciting things, very exciting things.

I also added folders for the the different cases, pll and oll, that
contain all of the images. So now I need to remove the code that says
that we need to add the img property...I also need to double check to
make sure that it is the 'img' property that we're looking for...