Finished modularizing code
I only have a few corner-cases type stuff to chase down before I merge
with the master branch, which include things like the length property
probably not being spelled correctly (which tends to be a problem in
these kinds of weird bugs) with the solves in the cases. I think I'll
also get rid of the oll-content-whatever folder in the app/ folder,
because that's been replaced by the content/ folder. Had a weird
mini-bug where the prob on cards was hidden whenever the probability
column was hidden, but that was because the probability class was used
to hide the column, and the card probability class was the same as the
column probability class, so that was cleared up very quickly once I
discovered it. I think I want to make the code more readable, and also
abstract it out in a more portable fashion, and actually make sure that
certain things, such as object references, are actually being passed
along, because otherwise, there's no point to the abstraction.

Also, I want to get the edit table button working again, but that's just
me. It's not the button that needs fixing, but its message.