Furthered PLL compatibility
I needed to make sure that the PLL algorithms, which are, on average,
longer than OLL algorithms, would show up properly on the flash cards.
For this, I needed to modify the flashCard.less file, as well as make
a separate directive for the PLL flashcards, versus the OLL flashcards.
The basic.less file was also altered in this change because I needed to
alter some of the ways in which the innerBox class is rendered in order
to make more room for the flash cards.

This update also solves compatibility issues with pllCases, which had
the property "len" in its solve1 and solve2 objects rather than the
property "length," which was easily rectified.

I also had to update server.js with code to fetch the node modules,
because it was unable to serve the dependencies that I had in the node
modules folder. This enables me to not only test while without wifi or
ethernet, but also to develop.