Tried to get it to build, not working, mostly fixed
I've fixed most of the bugs (the edit table button message, working on
further modularizing the code as it is, special functions just for the
$scope.$watch(Collection) callbacks, etc. I've only implemented one
function for the latter, but it's been working okay so far.

I also spotted a mistake in one of the images, which had to be
fixed/replaced with a new png. I believe it was Ab.

Something weird also happened with the less where everything was just
not rendering correctly, but I fixed it. I may make the less more
"professional" at some point in the near future.

I've also added functionality for processhtml and concat, for grunt.
That way, I can have multiple different builds for production,
development, distribution, etc. However, as stated in the title, I
haven't exactly gotten it to work, per se, yet.

I also had to fix a little thing with Jquery, because I switched to
using node modules instead of a remote file. They deprecated using
$(window).load() so I had to change that.