Added more to stats page
I added the count and the full range, essentiall the worst time minus
the best time, as well as tweaked the millisToStrings filter so that it
wouldn't turn the count into a string. I really ought to just pass in
the name of the thing so that I can just exclude count specifically, but
I will do that at some other point.

I also discovered that the clicking on the divider titles did not make
them accordion anymore so I had to fix that as well. Apparently you can
attach an event higher up in the html tree and it will just apply to
anything that is within that tree. That fixed that problem. I'm not
entirely sure how this problem came about, but it was probably when I
started using a directive instead of just code in the html.