Fixed about 17 tests, Added oll/pll-controller tests
I added tests for the oll and pll controllers. I also fixed all of the
flashcarddata tests that weren't working. It turned out to be a matter
of actually setting the factory's type (oll/pll) and also referencing
that type because I changed some of the structure of the objects in the
past few months and neglected to do something about that in the tests.

I only have 4 that don't work now, in addition to 4 tests that I skip.