Added more coverage, added deleteLastTime
I wanted a button that would delete the last time, so I added it, which
meant adding a function to timer.controller.js.

I also added more tests to the timer.spec.js file because apparently I
neglected to ever make deleteAll a test and I had to make a test for
deleteLastTime. We're officially over 60% for coverage in terms of
lines.