Adding support for different cubes + settings
I want to be able to have different settings that get saved for each
person, so I'm trying to get that set up. I got a little icon and it
should show up on the OLL page and the PLL page...maybe I should put it
on the home page as well.

I've been contemplating having a separate scrambler because running the
scrambler + timer + cases table + flashcards + controls is probably a
headache for both my browser and angular. I could probably also
streamline my process by using events instead of watches. Events have
been very useful in the past, I just need to get some kind of thing
together to do that.

The cases table also freaked out and had like 18px Arial font for a bit.
I fixed it though so it's back to normal.